Refactor packages to prepare it for db addition.
