+++ app/src/main/java/com/manugmoya/bicimadstations/ui/detail/DetailActivity.kt
@@ -4,10 +4,9 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import com.manugmoya.bicimadstations.R
 import com.manugmoya.bicimadstations.databinding.ActivityDetailBinding
-import com.manugmoya.bicimadstations.model.Station
+import com.manugmoya.bicimadstations.model.server.Station
 import com.manugmoya.bicimadstations.ui.common.getViewModel
 
 class DetailActivity : AppCompatActivity(){
