+++ app/src/main/java/com/manugmoya/bicimadstations/ui/detail/DetailViewModel.kt
@@ -3,8 +3,7 @@ package com.manugmoya.bicimadstations.ui.detail
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.manugmoya.bicimadstations.model.Station
+import com.manugmoya.bicimadstations.model.server.Station
 import com.manugmoya.bicimadstations.ui.common.Scope
 
 class DetailViewModel(private val station: Station) : ViewModel() ,Scope by Scope.Impl(){
